UserService/OcelotGateway: roles and cities finished

+++ UserService/Services/Cities/CitiesService.cs
@@ -4,15 +4,20 @@ using System.Linq;
 using System.Threading.Tasks;
 using UserService.Data;
 using UserService.Entities;
+using UserService.Exceptions;
 
 namespace UserService.Services.Cities
 {
     public class CitiesService : ICitiesService
     {
         private readonly ICityRepository _cityRepository;
-        public CitiesService(ICityRepository cityRepository)
+        private readonly IPersonalUserRepository _personalUserRepository;
+        private readonly ICorporationUserRepository _corporationUserRepository;
+        public CitiesService(ICityRepository cityRepository, ICorporationUserRepository corporationUserRepository, IPersonalUserRepository personalUserRepository)
         {
             _cityRepository = cityRepository;
+            _corporationUserRepository = corporationUserRepository;
+            _personalUserRepository = personalUserRepository;
         }
         public CityCreatedConfirmation CreateCity(City city)
         {
@@ -23,6 +28,16 @@ namespace UserService.Services.Cities
 
         public void DeleteCity(Guid cityId)
         {
+            List<PersonalUser> pUsers = _personalUserRepository.GetUsersWithCity(cityId);
+            if (pUsers.Count > 0)
+            {
+                throw new ReferentialConstraintViolationException("Value of city is referenced in another table");
+            }
+            List<Corporation> cUsers = _corporationUserRepository.GetUsersWithCity(cityId);
+            if (cUsers.Count > 0)
+            {
+                throw new ReferentialConstraintViolationException("Value of city is referenced in another table");
+            }
             _cityRepository.DeleteCity(cityId);
             _cityRepository.SaveChanges();
         }
